Final changes around nightfall
