+++ .stylelintrc.json
@@ -6,6 +6,7 @@
     "rules": {
       "prettier/prettier": true,
       "declaration-no-important": true,
-      "no-empty-source": null
+      "no-empty-source": null,
+      "selector-class-pattern": null
     }
   }
