+++ esbuild.js
@@ -2,10 +2,11 @@ import * as esbuild from "esbuild";
 
 const widgets = [
   'carousel',
-  'nightfall'
+  'nightfall',
+  'waterfall'
 ]
 
-await esbuild.build({
+const config = {
   entryPoints: [...widgets.map((widget) => `widgets/${widget}/widget.ts`)],
   bundle: true,
   outdir: "dist",
@@ -13,4 +14,14 @@ await esbuild.build({
     ".hbs": "text",
     ".css": "text",
   },
-});
+  minify: true
+}
+
+if (process.env.APP_ENV == 'development') {
+  config.minify = false;
+  config.sourcemap = 'inline';
+  const context = await esbuild.context(config);
+  await context.watch();
+} else {
+  await esbuild.build(config);
+}
