+++ widgets/nightfall/components/add-to-cart/base.styles.ts
@@ -1,6 +1,6 @@
 export default `
   .ugc-add-to-cart-button {
-    max-width: 300px !important;
+    max-width: 300px;
     text-transform: uppercase;
   }
 `;
