--- widgets/waterfall/components/cross-sellers/base.styles.ts
@@ -1,1 +1,0 @@
-export default ``;
