--- widgets/waterfall/components/cross-sellers/base.template.ts
@@ -1,3 +1,0 @@
-export default () => {
-  return ``;
-};
