+++ widgets/waterfall/components/expanded-tile/base.css
@@ -15,44 +15,31 @@
   position: absolute;
   top: 20px;
   right: 20px;
-  z-index: 1;
-  background: #fff;
-  border-radius: 50%;
-  width: 30px;
-  height: 30px;
-  display: flex;
-  flex-wrap: nowrap;
-  place-content: center center;
-  align-items: center;
 }
 
 .panel {
+  display: flex;
   position: relative;
-  background: #f4f4f4;
-  width: 600px;
+  background: #fff;
+  width: 85%;
   height: 100%;
-  margin: 30px auto;
-  display: block;
 }
 
 .panel-left {
+  min-width: 40%;
+  max-width: 40%;
   background: #fff;
   display: flex;
   place-content: center center;
   align-items: center;
-  width: 100%;
-  max-width: 100%;
-  min-width: 100%;
 }
 
 .panel-left img {
   width: 100%;
-  display: block;
 }
 
 .panel-right {
-  padding: 24px 44px;
-  background: #2c2d2c;
+  padding: 26px 40px;
 }
 
 .tile-arrows-left {
@@ -63,6 +50,7 @@
   left: -60px;
   border: 0;
   background: none;
+  z-index: 2;
 }
 
 .tile-arrows-right {
@@ -70,9 +58,10 @@
   width: 50px;
   height: 50px;
   top: 50%;
-  right: -50px;
+  right: -60px;
   border: 0;
   background: none;
+  z-index: 2;
 }
 
 .widget-icon {
@@ -96,7 +85,7 @@ div.image {
   align-items: center;
   justify-content: center;
   width: 100%;
-  margin-bottom: 0;
+  margin-bottom: 16px;
 }
 
 .image-wrapper-inner {
@@ -105,11 +94,6 @@ div.image {
   align-items: center;
   justify-content: center;
   width: 100%;
-}
-
-.image-wrapper-inner .image {
-  width: 100%;
-  display: block;
 }
 
 .user-info {
@@ -187,6 +171,11 @@ html {
   padding-right: 20px;
   text-align: left;
   width: 100%;
+}
+
+.ugc-inline-share-buttons {
+  margin: 16px;
+  text-align: right;
 }
 
 .tile-caption {
