+++ widgets/waterfall/css.variables.ts
